+++ R/colour-mapping.r
@@ -3,26 +3,26 @@
 #' Conveniently maps data values (numeric or factor/character) to colors
 #' according to a given palette, which can be provided in a variety of formats.
 #'
-#' \code{col_numeric} is a simple linear mapping from continuous numeric data
+#' `col_numeric` is a simple linear mapping from continuous numeric data
 #' to an interpolated palette.
 #'
 #' @param palette The colors or color function that values will be mapped to
 #' @param domain The possible values that can be mapped.
 #'
-#'   For \code{col_numeric} and \code{col_bin}, this can be a simple numeric
-#'   range (e.g. \code{c(0, 100)}); \code{col_quantile} needs representative
-#'   numeric data; and \code{col_factor} needs categorical data.
+#'   For `col_numeric` and `col_bin`, this can be a simple numeric
+#'   range (e.g. `c(0, 100)`); `col_quantile` needs representative
+#'   numeric data; and `col_factor` needs categorical data.
 #'
-#'   If \code{NULL}, then whenever the resulting color function is called, the
-#'   \code{x} value will represent the domain. This implies that if the function
+#'   If `NULL`, then whenever the resulting color function is called, the
+#'   `x` value will represent the domain. This implies that if the function
 #'   is invoked multiple times, the encoding between values and colors may not
-#'   be consistent; if consistency is needed, you must provide a non-\code{NULL}
+#'   be consistent; if consistency is needed, you must provide a non-`NULL`
 #'   domain.
-#' @param na.color The color to return for \code{NA} values. Note that
-#'   \code{na.color=NA} is valid.
+#' @param na.color The color to return for `NA` values. Note that
+#'   `na.color=NA` is valid.
 #'
-#' @return A function that takes a single parameter \code{x}; when called with a
-#'   vector of numbers (except for \code{col_factor}, which expects
+#' @return A function that takes a single parameter `x`; when called with a
+#'   vector of numbers (except for `col_factor`, which expects
 #'   factors/characters), #RRGGBB color strings are returned.
 #'
 #' @export
@@ -37,7 +37,7 @@ col_numeric <- function(palette, domain, na.color = "#808080") {
 
   pf <- safePaletteFunc(palette, na.color)
 
-  withColorAttr('numeric', list(na.color = na.color), function(x) {
+  withColorAttr("numeric", list(na.color = na.color), function(x) {
     if (length(x) == 0 || all(is.na(x))) {
       return(pf(x))
     }
@@ -45,8 +45,9 @@ col_numeric <- function(palette, domain, na.color = "#808080") {
     if (is.null(rng)) rng <- range(x, na.rm = TRUE)
 
     rescaled <- scales::rescale(x, from = rng)
-    if (any(rescaled < 0 | rescaled > 1, na.rm = TRUE))
+    if (any(rescaled < 0 | rescaled > 1, na.rm = TRUE)) {
       warning("Some values were outside the color scale and will be treated as NA")
+    }
     pf(rescaled)
   })
 }
@@ -81,15 +82,15 @@ getBins <- function(domain, x, bins, pretty) {
   }
 }
 
-#' @details \code{col_bin} also maps continuous numeric data, but performs
-#'   binning based on value (see the \code{\link[base]{cut}} function).
+#' @details `col_bin` also maps continuous numeric data, but performs
+#'   binning based on value (see the [base::cut()] function).
 #' @param bins Either a numeric vector of two or more unique cut points or a
 #'   single number (greater than or equal to 2) giving the number of intervals
 #'   into which the domain values are to be cut.
-#' @param pretty Whether to use the function \code{\link{pretty}()} to generate
-#'   the bins when the argument \code{bins} is a single number. When
-#'   \code{pretty = TRUE}, the actual number of bins may not be the number of
-#'   bins you specified. When \code{pretty = FALSE}, \code{\link{seq}()} is used
+#' @param pretty Whether to use the function [pretty()] to generate
+#'   the bins when the argument `bins` is a single number. When
+#'   `pretty = TRUE`, the actual number of bins may not be the number of
+#'   bins you specified. When `pretty = FALSE`, [seq()] is used
 #'   to generate the bins and the breaks may not be "pretty".
 #' @rdname col_numeric
 #' @export
@@ -100,39 +101,40 @@ col_bin <- function(palette, domain, bins = 7, pretty = TRUE, na.color = "#80808
     domain <- NULL
   }
   autobin <- is.null(domain) && length(bins) == 1
-  if (!is.null(domain))
+  if (!is.null(domain)) {
     bins <- getBins(domain, NULL, bins, pretty)
+  }
   numColors <- if (length(bins) == 1) bins else length(bins) - 1
   colorFunc <- col_factor(palette, domain = if (!autobin) 1:numColors, na.color = na.color)
-  pf = safePaletteFunc(palette, na.color)
+  pf <- safePaletteFunc(palette, na.color)
 
-  withColorAttr('bin', list(bins = bins, na.color = na.color), function(x) {
+  withColorAttr("bin", list(bins = bins, na.color = na.color), function(x) {
     if (length(x) == 0 || all(is.na(x))) {
       return(pf(x))
     }
     binsToUse <- getBins(domain, x, bins, pretty)
     ints <- cut(x, binsToUse, labels = FALSE, include.lowest = TRUE, right = FALSE)
-    if (any(is.na(x) != is.na(ints)))
+    if (any(is.na(x) != is.na(ints))) {
       warning("Some values were outside the color scale and will be treated as NA")
+    }
     colorFunc(ints)
   })
 }
 
-#' @details \code{col_quantile} similarly bins numeric data, but via the
-#'   \code{\link[stats]{quantile}} function.
+#' @details `col_quantile` similarly bins numeric data, but via the
+#'   [stats::quantile()] function.
 #' @param n Number of equal-size quantiles desired. For more precise control,
-#'   use the \code{probs} argument instead.
-#' @param probs See \code{\link[stats]{quantile}}. If provided, the \code{n}
+#'   use the `probs` argument instead.
+#' @param probs See [stats::quantile()]. If provided, the `n`
 #'   argument is ignored.
 #' @rdname col_numeric
 #' @export
 col_quantile <- function(palette, domain, n = 4,
-  probs = seq(0, 1, length.out = n + 1), na.color = "#808080") {
-
+                         probs = seq(0, 1, length.out = n + 1), na.color = "#808080") {
   if (!is.null(domain)) {
     bins <- stats::quantile(domain, probs, na.rm = TRUE, names = FALSE)
     return(withColorAttr(
-      'quantile', list(probs = probs, na.color = na.color),
+      "quantile", list(probs = probs, na.color = na.color),
       col_bin(palette, domain = NULL, bins = bins, na.color = na.color)
     ))
   }
@@ -141,11 +143,12 @@ col_quantile <- function(palette, domain, n = 4,
   # If you say probs = seq(0, 1, 0.25), which has length 5, does that map to 4 colors
   # or 5? 4, right?
   colorFunc <- col_factor(palette, domain = 1:(length(probs) - 1), na.color = na.color)
-  withColorAttr('quantile', list(probs = probs, na.color = na.color), function(x) {
+  withColorAttr("quantile", list(probs = probs, na.color = na.color), function(x) {
     binsToUse <- stats::quantile(x, probs, na.rm = TRUE, names = FALSE)
     ints <- cut(x, binsToUse, labels = FALSE, include.lowest = TRUE, right = FALSE)
-    if (any(is.na(x) != is.na(ints)))
+    if (any(is.na(x) != is.na(ints))) {
       warning("Some values were outside the color scale and will be treated as NA")
+    }
     colorFunc(ints)
   })
 }
@@ -165,8 +168,9 @@ calcLevels <- function(x, ordered) {
 }
 
 getLevels <- function(domain, x, lvls, ordered) {
-  if (!is.null(lvls))
+  if (!is.null(lvls)) {
     return(lvls)
+  }
 
   if (!is.null(domain)) {
     return(calcLevels(domain, ordered))
@@ -177,17 +181,17 @@ getLevels <- function(domain, x, lvls, ordered) {
   }
 }
 
-#' @details \code{col_factor} maps factors to colors. If the palette is
+#' @details `col_factor` maps factors to colors. If the palette is
 #'   discrete and has a different number of colors than the number of factors,
 #'   interpolation is used.
 #' @param levels An alternate way of specifying levels; if specified, domain is
 #'   ignored
-#' @param ordered If \code{TRUE} and \code{domain} needs to be coerced to a
+#' @param ordered If `TRUE` and `domain` needs to be coerced to a
 #'   factor, treat it as already in the correct order
 #' @rdname col_numeric
 #' @export
 col_factor <- function(palette, domain, levels = NULL, ordered = FALSE,
-  na.color = "#808080") {
+                       na.color = "#808080") {
 
   # domain usually needs to be explicitly provided (even if NULL) but not if
   # levels are specified
@@ -203,7 +207,7 @@ col_factor <- function(palette, domain, levels = NULL, ordered = FALSE,
   hasFixedLevels <- is.null(lvls)
   pf <- safePaletteFunc(palette, na.color)
 
-  withColorAttr('factor', list(na.color = na.color), function(x) {
+  withColorAttr("factor", list(na.color = na.color), function(x) {
     if (length(x) == 0 || all(is.na(x))) {
       return(pf(x))
     }
@@ -229,11 +233,11 @@ col_factor <- function(palette, domain, levels = NULL, ordered = FALSE,
   })
 }
 
-#' @details The \code{palette} argument can be any of the following:
+#' @details The `palette` argument can be any of the following:
 #' \enumerate{
-#'   \item{A character vector of RGB or named colors. Examples: \code{palette()}, \code{c("#000000", "#0000FF", "#FFFFFF")}, \code{topo.colors(10)}}
-#'   \item{The name of an RColorBrewer palette, e.g. \code{"BuPu"} or \code{"Greens"}.}
-#'   \item{A function that receives a single value between 0 and 1 and returns a color. Examples: \code{colorRamp(c("#000000", "#FFFFFF"), interpolate="spline")}.}
+#'   \item{A character vector of RGB or named colors. Examples: `palette()`, `c("#000000", "#0000FF", "#FFFFFF")`, `topo.colors(10)`}
+#'   \item{The name of an RColorBrewer palette, e.g. `"BuPu"` or `"Greens"`.}
+#'   \item{A function that receives a single value between 0 and 1 and returns a color. Examples: `colorRamp(c("#000000", "#FFFFFF"), interpolate="spline")`.}
 #' }
 #' @examples
 #' pal <- col_bin("Greens", domain = 0:100)
@@ -259,7 +263,8 @@ NULL
 
 safePaletteFunc <- function(pal, na.color) {
   filterRange(
-    filterNA(na.color = na.color,
+    filterNA(
+      na.color = na.color,
       filterZeroLength(
         filterRGB(
           toPaletteFunc(pal)
@@ -278,7 +283,7 @@ toPaletteFunc <- function(pal) {
 toPaletteFunc.character <- function(pal) {
   if (length(pal) == 1 && pal %in% row.names(RColorBrewer::brewer.pal.info)) {
     return(colour_ramp(
-      RColorBrewer::brewer.pal(RColorBrewer::brewer.pal.info[pal, 'maxcolors'], pal)
+      RColorBrewer::brewer.pal(RColorBrewer::brewer.pal.info[pal, "maxcolors"], pal)
     ))
   }
 
@@ -345,33 +350,36 @@ filterRange <- function(f) {
 
 #' Fast color interpolation
 #'
-#' Returns a function that maps the interval [0,1] to a set of colors.
+#' Returns a function that maps the interval \[0,1] to a set of colors.
 #' Interpolation is performed in the CIELAB color space. Similar to
 #' \code{\link[grDevices]{colorRamp}(space = 'Lab')}, but hundreds of
-#' times faster, and provides results in \code{"#RRGGBB"} (or
-#' \code{"#RRGGBBAA"}) character form instead of RGB color matrices.
+#' times faster, and provides results in `"#RRGGBB"` (or
+#' `"#RRGGBBAA"`) character form instead of RGB color matrices.
 #'
 #' @param colors Colors to interpolate; must be a valid argument to
-#'   \code{\link[grDevices]{col2rgb}}. This can be a character vector of
-#'   \code{"#RRGGBB"} or  \code{"#RRGGBBAA"}, color names from
-#'   \code{\link[grDevices]{colors}}, or a positive integer that indexes into
-#'   \code{\link[grDevices]{palette}()}.
-#' @param na.color The color to map to \code{NA} values (for example,
-#'   \code{"#606060"} for dark grey, or \code{"#00000000"} for transparent) and
-#'   values outside of [0,1]. Can itself by \code{NA}, which will simply cause
-#'   an \code{NA} to be inserted into the output.
-#' @param alpha Whether to include alpha channels in interpolation; otherwise,
-#'   any alpha information will be discarded. If \code{TRUE} then the returned
-#'   function will provide colors in \code{"#RRGGBBAA"} format instead of
-#'   \code{"#RRGGBB"}.
+#'   [grDevices::col2rgb()]. This can be a character vector of
+#'   `"#RRGGBB"` or  `"#RRGGBBAA"`, color names from
+#'   [grDevices::colors()], or a positive integer that indexes into
+#'   [grDevices::palette()].
+#' @param na.color The color to map to `NA` values (for example,
+#'   `"#606060"` for dark grey, or `"#00000000"` for transparent) and
+#'   values outside of \[0,1]. Can itself by `NA`, which will simply cause
+#'   an `NA` to be inserted into the output.
+#' @param alpha Whether to include alpha transparency channels in interpolation.
+#'   If `TRUE` then the alpha information is included in the interpolation.
+#'   The returned colors will be provided in `"#RRGGBBAA"` format when needed,
+#'   i.e., in cases where the color is not fully opaque, so that the `"AA"`
+#'   part is not equal to `"FF"`. Fully opaque colors will be returned in
+#'   `"#RRGGBB"` format. If `FALSE`, the alpha information is discared
+#'   before interpolation and colors are always returned as `"#RRGGBB"`.
 #'
 #' @return A function that takes a numeric vector and returns a character vector
 #'   of the same length with RGB or RGBA hex colors.
 #'
-#' @seealso \link[grDevices]{colorRamp}
+#' @seealso \code{\link[grDevices]{colorRamp}}
 #'
 #' @export
-colour_ramp <- function(colors, na.color = NA, alpha = FALSE) {
+colour_ramp <- function(colors, na.color = NA, alpha = TRUE) {
   if (length(colors) == 0) {
     stop("Must provide at least one color to create a color ramp")
   }