+++ tests/testthat/test-bounds.r
@@ -31,29 +31,30 @@ test_that("zero range inputs return mid range", {
 test_that("censor and squish ignore infinite values", {
   expect_equal(squish(c(1, Inf)), c(1, Inf))
   expect_equal(censor(c(1, Inf)), c(1, Inf))
-
-
 })
 
 test_that("scaling is possible with dates and times", {
   dates <- as.Date(c("2010-01-01", "2010-01-03", "2010-01-05", "2010-01-07"))
-  expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0,1,1/3))
+  expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0, 1, 1 / 3))
   expect_equal(rescale_mid(dates, mid = dates[3])[3], 0.5)
 
-  dates <- as.POSIXct(c("2010-01-01 01:40:40",
-                        "2010-01-01 03:40:40",
-                        "2010-01-01 05:40:40",
-                        "2010-01-01 07:40:40"))
-  expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0, 1, 1/3))
+  dates <- as.POSIXct(c(
+    "2010-01-01 01:40:40",
+    "2010-01-01 03:40:40",
+    "2010-01-01 05:40:40",
+    "2010-01-01 07:40:40"
+  ))
+  expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0, 1, 1 / 3))
   expect_equal(rescale_mid(dates, mid = dates[3])[3], 0.5)
-
 })
 
 test_that("scaling is possible with integer64 data", {
+  skip_if_not_installed("bit64")
   x <- bit64::as.integer64(2^60) + c(0:3)
   expect_equal(
     rescale_mid(x, mid = bit64::as.integer64(2^60) + 1),
-    c(0.25, 0.5, 0.75, 1))
+    c(0.25, 0.5, 0.75, 1)
+  )
 })
 
 test_that("scaling is possible with NULL values", {