+++ tests/testthat/test-formatter.r
@@ -17,21 +17,27 @@ test_that("scientific format shows specific sig figs", {
 })
 
 test_that("wrap correctly wraps long lines", {
-  expect_equal(wrap_format(10)("this is a long line"),
-               "this is a\nlong line")
-  expect_equal(wrap_format(10)(c("this is a long line", "this is another long line")),
-               c("this is a\nlong line", "this is\nanother\nlong line"))
-  expect_equal(wrap_format(10)("a_very_long_word_without_spaces"),
-               "a_very_long_word_without_spaces")
+  expect_equal(
+    wrap_format(10)("this is a long line"),
+    "this is a\nlong line"
+  )
+  expect_equal(
+    wrap_format(10)(c("this is a long line", "this is another long line")),
+    c("this is a\nlong line", "this is\nanother\nlong line")
+  )
+  expect_equal(
+    wrap_format(10)("a_very_long_word_without_spaces"),
+    "a_very_long_word_without_spaces"
+  )
   expect_equal(wrap_format(10)("short line"), "short\nline")
   expect_equal(wrap_format(15)("short line"), "short line")
 })
 
 test_that("ordinal format", {
-  expect_equal(ordinal(1),  "1st")
-  expect_equal(ordinal(2),  "2nd")
-  expect_equal(ordinal(3),  "3rd")
-  expect_equal(ordinal(4),  "4th")
+  expect_equal(ordinal(1), "1st")
+  expect_equal(ordinal(2), "2nd")
+  expect_equal(ordinal(3), "3rd")
+  expect_equal(ordinal(4), "4th")
   expect_equal(ordinal(11), "11th")
   expect_equal(ordinal(12), "12th")
   expect_equal(ordinal(21), "21st")
@@ -43,7 +49,7 @@ test_that("ordinal format maintains order", {
 
 test_that("formatters don't add extra spaces", {
   has_space <- function(x) any(grepl("\\s", x))
-  x <- 10 ^ c(-1, 0, 1, 3, 6, 9)
+  x <- 10^c(-1, 0, 1, 3, 6, 9)
 
   expect_false(has_space(comma(x)))
   expect_false(has_space(dollar(x)))